refactor (#1605)
* Tidying up GSOC refactorization

Will end hopefully in a simpler structure.

* update
